fix song length being incorrect
I forgot the float cast haha so silly and goofy of me

+++ Assets/WebAudioClip/WebAudioClip.cs
@@ -26,7 +26,7 @@ public class WebAudioClip : IDisposable
 
         // Creates clip in the browser via JS
         clipId = WebAudioController.NewClip(channels, lengthSamples, frequency);
-        Length = lengthSamples / sampleRate;
+        Length = (float)lengthSamples / sampleRate;
 #endif
     }
